refactor: 调整平台禁用写法和 resolver 的位置
